requirements.txt
Django==2.2
django-environ==0.4.5
Pillow==7.0.0
python-slugify==4.0.0
